Extract as vite plugin
